package setting / rename some class / add annotation
