--- app/src/main/java/com/fancertification/www/ChannelData.kt
@@ -1,8 +1,0 @@
-package com.fancertification.www
-
-import java.io.Serializable
-
-data class ChannelData(
-    var data:SearchData, var viewCount:Long=0, var subscriberCount:Int=0, var videoCount: Int =0
-): Serializable
-
