+++ app/src/main/java/com/fancertification/www/FragmentHandler.kt
@@ -1,9 +1,10 @@
 package com.fancertification.www
-import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
+import com.fancertification.www.bookmark.BookmarkFragment
+import com.fancertification.www.search.SearchFragment
+import com.fancertification.www.setting.SettingFragment
 
 class FragmentHandler(fragmentActivity: FragmentActivity) :
     FragmentStateAdapter(fragmentActivity) {
@@ -16,7 +17,7 @@ class FragmentHandler(fragmentActivity: FragmentActivity) :
 
     override fun createFragment(position: Int): Fragment {  //뷰페이저 페이지 전환
         return when (position) {
-            0 -> BookmarkChannel()
+            0 -> BookmarkFragment()
             1 -> SearchFragment()
             2 -> SettingFragment()
             else -> SearchFragment()
