--- app/src/main/java/com/fancertification/www/SettingFragment.kt
@@ -1,11 +1,0 @@
-package com.fancertification.www
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-class SettingFragment : Fragment(R.layout.fragment_setting) {
-
-
-}
