+++ app/src/main/java/com/fancertification/www/setting/SettingFragment.kt
@@ -1,0 +1,9 @@
+package com.fancertification.www.setting
+
+import androidx.fragment.app.Fragment
+import com.fancertification.www.R
+
+class SettingFragment : Fragment(R.layout.fragment_setting) {
+
+
+}
