+++ app/src/main/res/layout/fragment_bookmark.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="@color/white"
-    tools:context=".BookmarkChannel">
+    tools:context=".bookmark.BookmarkFragment">
     <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 
         android:id="@+id/swiperefresh"
