+++ app/src/main/res/layout/fragment_setting.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".SettingFragment">
+    tools:context=".setting.SettingFragment">
 
 
     <TextView
