+++ app/src/main/res/layout/search_adapter_item.xml
